# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DocArchiv",
    "description": "Leichtgewichtiger Dokumenten-Katalog mit Volltextsuche und Nextcloud-Linkout.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/correspondents": {
      "get": {
        "tags": [
          "correspondents"
        ],
        "summary": "List Correspondents",
        "description": "Liefert alle Korrespondenten inklusive Dokumentanzahl.",
        "operationId": "list_correspondents_api_correspondents_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CorrespondentResponse"
                  },
                  "type": "array",
                  "title": "Response List Correspondents Api Correspondents Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "correspondents"
        ],
        "summary": "Create Correspondent",
        "description": "Legt einen neuen Korrespondenten an.",
        "operationId": "create_correspondent_api_correspondents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorrespondentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CorrespondentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/correspondents/{correspondent_id}": {
      "get": {
        "tags": [
          "correspondents"
        ],
        "summary": "Get Correspondent",
        "description": "Liefert einen einzelnen Korrespondenten anhand seiner ID.",
        "operationId": "get_correspondent_api_correspondents__correspondent_id__get",
        "parameters": [
          {
            "name": "correspondent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Correspondent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CorrespondentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "correspondents"
        ],
        "summary": "Update Correspondent",
        "description": "Aktualisiert einen vorhandenen Korrespondenten teilweise.",
        "operationId": "update_correspondent_api_correspondents__correspondent_id__patch",
        "parameters": [
          {
            "name": "correspondent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Correspondent Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorrespondentUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CorrespondentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "correspondents"
        ],
        "summary": "Delete Correspondent",
        "description": "Loescht einen Korrespondenten.",
        "operationId": "delete_correspondent_api_correspondents__correspondent_id__delete",
        "parameters": [
          {
            "name": "correspondent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Correspondent Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CorrespondentCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CorrespondentCreate",
        "description": "Request-Schema zum Anlegen eines Korrespondenten."
      },
      "CorrespondentResponse": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "document_count": {
            "type": "integer",
            "title": "Document Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "name",
          "id"
        ],
        "title": "CorrespondentResponse",
        "description": "Response-Schema fuer Korrespondenten."
      },
      "CorrespondentUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "type": "object",
        "title": "CorrespondentUpdate",
        "description": "Request-Schema fuer teilweise Korrespondent-Updates."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}